moved constants.py
