+++ docs/api/physics/constants.md
@@ -1,31 +1,31 @@
 # Constants and Material Properties
 
 ## Physical Constants
-::: fdtdx.core.physics.constants.c
-::: fdtdx.core.physics.constants.mu0
-::: fdtdx.core.physics.constants.eps0
-::: fdtdx.core.physics.constants.eta0
-::: fdtdx.core.physics.constants.wavelength_to_period
+::: fdtdx.constants.c
+::: fdtdx.constants.mu0
+::: fdtdx.constants.eps0
+::: fdtdx.constants.eta0
+::: fdtdx.constants.wavelength_to_period
 
 ## Material Properties
-::: fdtdx.core.physics.constants.relative_permittivity_air
-::: fdtdx.core.physics.constants.relative_permittivity_substrate
-::: fdtdx.core.physics.constants.relative_permittivity_polymer
-::: fdtdx.core.physics.constants.relative_permittivity_silicon
-::: fdtdx.core.physics.constants.relative_permittivity_silica
-::: fdtdx.core.physics.constants.relative_permittivity_SZ_2080
-::: fdtdx.core.physics.constants.relative_permittivity_ma_N_1400_series
-::: fdtdx.core.physics.constants.relative_permittivity_bacteria
-::: fdtdx.core.physics.constants.relative_permittivity_water
-::: fdtdx.core.physics.constants.relative_permittivity_fused_silica
-::: fdtdx.core.physics.constants.relative_permittivity_coated_silica
-::: fdtdx.core.physics.constants.relative_permittivity_resin
-::: fdtdx.core.physics.constants.relative_permittivity_ormo_prime
+::: fdtdx.constants.relative_permittivity_air
+::: fdtdx.constants.relative_permittivity_substrate
+::: fdtdx.constants.relative_permittivity_polymer
+::: fdtdx.constants.relative_permittivity_silicon
+::: fdtdx.constants.relative_permittivity_silica
+::: fdtdx.constants.relative_permittivity_SZ_2080
+::: fdtdx.constants.relative_permittivity_ma_N_1400_series
+::: fdtdx.constants.relative_permittivity_bacteria
+::: fdtdx.constants.relative_permittivity_water
+::: fdtdx.constants.relative_permittivity_fused_silica
+::: fdtdx.constants.relative_permittivity_coated_silica
+::: fdtdx.constants.relative_permittivity_resin
+::: fdtdx.constants.relative_permittivity_ormo_prime
 
 ## Material Configurations
-::: fdtdx.core.physics.constants.silicon_permittivity_config
-::: fdtdx.core.physics.constants.standard_permittivity_config
-::: fdtdx.core.physics.constants.higher_permittivity_config
-::: fdtdx.core.physics.constants.silica_permittivity_config
-::: fdtdx.core.physics.constants.multi_material_permittivity_config
-::: fdtdx.core.physics.constants.multi_material_qd_permittivity_config
+::: fdtdx.constants.silicon_permittivity_config
+::: fdtdx.constants.standard_permittivity_config
+::: fdtdx.constants.higher_permittivity_config
+::: fdtdx.constants.silica_permittivity_config
+::: fdtdx.constants.multi_material_permittivity_config
+::: fdtdx.constants.multi_material_qd_permittivity_config
