+++ examples/optimize_ceviche_corner.py
@@ -17,7 +17,7 @@ from fdtdx.constraints.module import (
     StandardToPlusOneMinusOneRange,
 )
 from fdtdx.config import GradientConfig, SimulationConfig
-from fdtdx.core.physics import constants
+from fdtdx import constants
 from fdtdx.core.physics.losses import metric_efficiency
 from fdtdx.fdtd.backward import full_backward
 from fdtdx.fdtd.fdtd import reversible_fdtd
