+++ examples/optimize_waveguide_stitching.py
@@ -11,7 +11,7 @@ from fdtdx.constraints.discrete import BOTTOM_Z_PADDING_CONFIG_REPEAT, BinaryMed
 from fdtdx.constraints.mapping import ConstraintMapping
 from fdtdx.constraints.module import ClosestIndex, IndicesToInversePermittivities, StandardToInversePermittivityRange
 from fdtdx.config import GradientConfig, SimulationConfig
-from fdtdx.core.physics import constants
+from fdtdx import constants
 from fdtdx.core.physics.losses import metric_efficiency
 from fdtdx.fdtd.backward import full_backward
 from fdtdx.fdtd.fdtd import reversible_fdtd
