+++ examples/speed_benchmark.py
@@ -16,7 +16,7 @@ from loguru import logger
 
 from fdtdx.constraints.mapping import ConstraintMapping
 from fdtdx.constraints.module import ClosestIndex, IndicesToInversePermittivities, StandardToInversePermittivityRange
-from fdtdx.core.physics import constants
+from fdtdx import constants
 from fdtdx.config import GradientConfig, SimulationConfig
 from fdtdx.core.physics.losses import metric_efficiency
 from fdtdx.core.plotting import colors
